requirements.txt
Django==1.7
django-appconf==1.0.1
django-geoposition==0.2.2
django-taggit==0.17.1
mysql==0.0.1
MySQL-python==1.2.5
Pillow==2.9.0
python-memcached==1.57
six==1.9.0
sorl-thumbnail==12.3
wheel==0.24.0
